// src/types/featureType.ts
export default interface IFeatureType {
    elevationAvg?: number;
    elevationMin?: number;
    elevationMax?: number;
    elevationStd?: number;
    climateAvg?: number;
    climateMin?: number;
    climateMax?: number;
    climatePrecip?: number;
    climateHumidity?: number;
    climateWindSpd?: number;
    vegetationCover?: number;
    vegetationHeight?: number;
    vegetationDeparture?: number;
}